+++ src/Structures/CPalFile.cpp
@@ -31,7 +31,7 @@ int CPalFile::CreateInstance(char* buffer)
 
 CPalFile* CPalFile::GetInstance(int index)
 {
-    return index >= 0 && index < Instances.size() ? &Instances[index] : nullptr;
+    return index >= 0 && index < (int)Instances.size() ? &Instances[index] : nullptr;
 }
 
 void CPalFile::Decode()
